pyproject.toml
[tool.poetry]
name = "order-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "order_backend" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.0"
gunicorn = "^21.2.0"
uvicorn = "^0.23.2"
mongoengine = "^0.27.0"
pydantic-settings = "^2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
